requirements.txt
streamlit==1.50.0google-genaipython-dotenv